requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
networkx>=3.0
pydantic>=2.0
pandas>=2.0
matplotlib>=3.7
python-dotenv>=1.0
flask>=3.0.0
